+++ src/GafferImageTest/ContextSanitiser.cpp
@@ -48,21 +48,6 @@ using namespace Gaffer;
 using namespace GafferImage;
 using namespace GafferImageTest;
 
-/// \todo Remove
-#ifndef IECORE_INTERNEDSTRING_WITH_TBB_HASHER
-
-namespace IECore
-{
-
-size_t tbb_hasher( const InternedString &s )
-{
-	return tbb::tbb_hasher( s.string() );
-}
-
-} // namespace IECore
-
-#endif
-
 ContextSanitiser::ContextSanitiser()
 {
 }